Default logging and metrics aspects
